use 'zimmer-server' as the next server for cobbler
orchestra adds `hostname` to /etc/hosts as 127.0.1.1 (via
orchestra-provisioning-server.preinst).  if 'zimmer' is both 
cobbler/server_and_next_server and the hostname, then cobbler's
attempt to resolve an ip_address when rendering files in
/var/lib/tftpboot/pxelinux.cfg/ through pxegen.py will come up
with 127.0.1.1, and thusly a url for the preseed file on the
linux command line like:
  url=http://127.0.1.1/cblr/svc/op/ks/system/odev-node01

Here, we add a host named 'zimmer-server' that is an alias for
zimmer, and in the build of zimmer populate that with 127.0.1.1.
However, at runtime, zimmer-server will be resolvable to the correct IP
address.  And it is at system add point that it is rendered to an IP.