Name this initial version for metadata service (2012-03-01).
This sets a specific name for this version of the metadata service.

By doing this, we can, in the future let 'latest' refer to the generic
idea of "the latest version", but we can support clients that want a given
version of the content.

The version of cloud-init in 12.04 will request:
   <url>/metadata/2012-03-01/<field>

And there-by identify the behavior of the server that it is expecting, and
ensuring that it will not be broken as long as servers implement that behavior
under that version.

Regarding the format of the version (YYYY-MM-DD) this is what amazon uses
in their metadata service, I'm by no means stuck to it.