zimmer-build: add ubuntu-orchestra-logging-server package
installed nodes via cobbler support rsyslog back to zimmer, but
we did not have the ubuntu-orchestra-logging-server package installed,
so we were not getting logs from the systems (at least during install).

orchestra will be replaced at some point, so this will have to change,
but automatically getting syslogs of install back to zimmer is worth it.