vdenv/zimmer-build/ud-build.txt: make 'apt_proxy' modify cobbler proxy
now, by setting the 'apt_proxy' value in zimmer-build/ud-build.txt, you
will get the zimmer system configured to use that proxy.  This is to address
the fact that we can't set this in a more direct way (LP: #914202).

By setting 'apt_proxy' to 'your-local-proxy:3128', now installs that happen
from that zimmer-build will get that proxy selected.