split image updating out from import-ephemerals
Thsi splits the image updating out into its own script:
   maas-cloudimg2ephemeral 
The reasoning for this is to make it more easy to be removed.
The update process is slow, and involves downloading kernel and other
packages and installing.  We will probably end up creating this
clean image elsewhere and making it available for download.