vdenv: ud-build default to importing isos of same release of zimmer
Previously, we update /etc/maas/import_isos to only import 'precise'
as a static string.  This just changes that to import whatever the
zimmer system is (ie, if its quantal, import quantal).  This can
easily be edited later and 'maas-import-isos' re-run.

Also note, that at present this does not affect maas-import-ephemerals.
That has its own default, which is still 'precise'.