make enlistment use ephemeral environment via enlistment metadata service
The overall goal of these changes is to get enlistment done in the
ephemeral environment rather than from the netboot ISO.  This is done by
adding an enlistment datasource for cloud-init.  

The "default" profile (which currently does enlistment) will be passed a
'url=' value that points to the rendered cloud-config content instructing
cloud-init to use the maas provided metadata service.

Then, the same path is taken by cloud-init as is done in the commissioning
environment, except for the fact that the metadata service that provides
the user-data will not require oauth.

Both cobbler and non-cobbler paths have been updated.