fix creation of cobbler profiles and improve "NO_COBBLER"
NO_COBBLER logic looked funny.  This makes the "cobbler only" code
more obvious by surrounding it with positive logic rather than negative
logic, and by ending the blocks with ending comment:

if USE_COBBLER; then
  # this code is cobbler only
fi

additionally, we fix a regression introduced here by the
'maas-import-pxe-files' changes:
-        kernel="$fpfinal_d/kernel"
-        initrd="$fpfinal_d/initrd"
+        kernel="$fpfinal_d/linux"
+        initrd="$fpfinal_d/initrd.gz"

The cobber code was expecting the name 'kernel' and 'initrd', but this
was arbitrarily changed to 'linux' and 'initrd.gz'.

Finally, we change the cobbler_add_update to check and make sure
that the add or update resulted in a profile.  Note, it does not check
that the update actually occurred, though.