if enlistment fails, offer login for debug
This sets a username password of 'ubuntu:ubuntu' if the user enlistment
process fails, and prints a message to the screen about the ability
to login with those credentials.

Upon login, a file is touched in /tmp , and if that file is present after
the script sleeps for 60 seconds, then the reboot is aborted.

In the default images, ssh via password auth is disabled, so unless that
is *enabled*, there is no concern of remotely coming in as that user.