Create a table named Tag, and give it a many-to-many relationship with nodes.
There is still a question about whether it is many-to-many in which way (does a node
have many tags, or do tags have many nodes). I think Juju is going to
do the queries as 'give me the nodes matching this tag', but the Web UI is
normally going to display 'give me the tags for this node'. Oh well.