When a tag gets a new definition, we need to repopulate the database with the new content.
This triggers on 'put' updating a field. I wonder if there is
another api for creating a Tag that isn't getting covered (namely
put() of new content into the database, but I guess the 'update'
code will already 404 if the object doesn't already exist?)

Anyway, I think this covers our cases for when we need to
rebuild the content. Tag.populate_nodes also gives us an
internal api where we can trigger an async update. [Flag
the tag as needing refreshing, and an async process can
poll for them and rebuild the node_tags table.