Pass ignore_result=True to all current celery tasks.
With celery, you are meant to fire off a job, and then grab the
result object and wait for it. The design of all current jobs is
that they just trigger and you never wait for the result.

If you don't ignore the results, then the queues build up forever
and stay alive because they wait for someone to read the result.