Small performance tweaks for the 'nodes/' and 'tags' pages.
We list a lot of nodes on those pages (we are looking to do 50, but
right now we do all of them.) However, we were previously doing
a unique query for every node to lookup that node's mac addresses.
We can use 'prefetch_all' so that Django does one query and caches it.
However, we have to switch from 'mac_addresses.reverse' to
'mac_addresses.all reversed' in the template. This is because
Django only caches the 'all' form. However, the template code
supports the 'reversed' syntax which essentially does the
reverse part in python.
Which works fine for what we want.