Performance tweak the nodegroup.list_nodes api.
Tell Django that we only care about the 'system_id' column,
which allows it to do its work in a much more efficient manner.
The time for the api call drops from 1.5s down to just 0.4s.
This is most likely because of reading all of the hardware_details
content, that we now get to skip.


Also, allow 'admin' users to query these URLs, as they can
see everything anyway. It means we can manually issue requests
against the api, which makes it much more reasonable to
performance test.