change poll based solution to upstart event based
The preivous solution relied on a backgrounded script to poll for status
and then poweroff.  This solution uses upstart events.

User scripts are executed by the 'cloud-final' job of cloud-init.  This
writes a job that starts on 'stopped cloud-final'.  The result is that 
upon cloud-final exiting, the poweroff job will be called.