1000 nodes seems to be a sweet spot for processing.
We don't quite hit 100% CPU utilization (80%), however at 5000 nodes,
the size of the request starts to be noticeable (120MB vs 24MB).
At 100 nodes, it takes 14s to process one tag, at 1000 and 5000 nodes
it takes 12s. So there isn't a wall-clock time win, and we have
a better upper bound on memory and network batch size.