Change how nodes/?op=accept checks that the system_ids are valid.
The existing check always loaded all nodes, and *if* there were
more nodes in the database than the list that was passed,
then it would assume that the ids passed were valid.
However, this both always loads all nodes, and allows you to
pass invalid system_ids, as long as you don't pass enough
to exceed the number of nodes in the db.
There isn't a security hole here, because you'll just get
PermissionDenied later, but you won't get the nice 'no such node'
error.
However, there is a performance fix, in that if they don't pass
any system_ids, we don't need to load the whole db to do a check
that doesn't do anything. And if they do pass system_ids, we
can load just the nodes requested, rather than the whole db again.