Clean up the actual exception handling.
In testing this with live data, it turns out that urllib2.open
raises an exception httplib.CONFLICT, rather than returning
the file-like object with a non OK status.
So fix the client to properly just log an info error.
Also, raising an exception generates a big traceback, which isn't
helpful, as there isn't a code error. So we just return
HttpResponse(status=httplib.CONFLICT) and things work nicer.